Token indices fixed, generic corpus class
Dedicated document class for sentence level documents
token indices fixed to match java index conventions (upper bound not
inclusive)
corpus is generic class wrt. document type
properties file for dataset paths

+++ src/Factors/Factor.java
@@ -13,7 +13,7 @@ public class Factor {
 		Log.off();
 	}
 
-	private static int factorIDIndex;
+	private static long factorIDIndex;
 	private Template template;
 	private FactorID factorID;
 	private Vector features;
